Improve search for MNAR cutoff
Instead of visual inspection of sorted heatmaps, we use the inflection
point of stacked probability densities of the proportion of missing
values/complete cases as a function of intensity. We use the cutoff
returned by the Control condition only for downstream analysis, as the
variability is lower and should provide a more accurate cutoff.